requirements.txt
Django==5.0.3
djangorestframework==3.15.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-django==4.8.0
pytest-drf==1.1.3
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0